+++ CHANGELOG.markdown
@@ -2,6 +2,22 @@
 
 
 
+## v0.5.1 (2011/01/30) ##
+
+### API changes ###
+
+ - made `SignalBinding` constructor private. (issue #15)
+ - changed params order on `SignalBinding` constructor.
+ - removed `signals.isDef()`. (issue #14)
+ 
+### Other ###
+
+ - added JSLint to the build process. (issue #12)
+ - validated source code using JSLint. (issue #13)
+ - improved docs.
+
+
+
 ## v0.5 (2010/12/03) ##
 
 ### API changes ###
@@ -11,12 +27,12 @@
    - `Signal.dispose()` (issue #6)
    - `signals.VERSION`
    - `signals.isDef()`
-    
+
  - Removed:
-  - `SignalBinding.listener` (issue #3)
-  
+   - `SignalBinding.listener` (issue #3)
+
  - Renamed:
-  - `SignalBinding.listenerScope` -> `SignalBinding.context` (issue #4)
+   - `SignalBinding.listenerScope` -> `SignalBinding.context` (issue #4)
 
 ### Fixes ###